refactor: Declaration global variable
